pyproject.toml
[project]
name = "pydantic-ai-code"
version = "0.1.0"
description = "A professional dual-mode AI assistant with async architecture for conversational chat and text rephrasing"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic-ai[openai]>=1.7.0",
    "python-dotenv>=1.0.0",
    "redis>=7.0.1",
    "ruff>=0.14.2",
    "flask>=3.0.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
